Update callback function parameter types to use `uint16_t` instead of `uint8_t`.
- Changed the parameter type of `toggleBacklight` function in `Callbacks.ino` from `uint8_t` to `uint16_t`.
- Changed the parameter type of `callback` function in `IntFloatValues.ino` from `uint8_t` to `uint16_t`.
- Changed the parameter types of `colorsCallback` and `numsCallback` functions in `List.ino` from `uint8_t` to `uint16_t`.
- Updated the typedefs in Constants.h for callback functions with integer parameters.
- Updated the return types of getItemIndex() functions in ItemList.h and ItemProgress.h classes from uint8_t to uint16_t.

This commit updates callback function parameter types and related definitions.

+++ examples/Callbacks/Callbacks.ino
@@ -23,7 +23,7 @@
 #define CLEAR 46     // NUMPAD .
 
 // Declare the call back function
-void toggleBacklight(uint8_t isOn);
+void toggleBacklight(uint16_t isOn);
 
 // prettier-ignore
 MAIN_MENU(
@@ -61,4 +61,4 @@ void loop() {
 /**
  * Define callback
  */
-void toggleBacklight(uint8_t isOn) { menu.setBacklight(isOn); }
+void toggleBacklight(uint16_t isOn) { menu.setBacklight(isOn); }
